+++ lib/cb/clients/data_list.rb
@@ -8,19 +8,30 @@
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
+require_relative 'base'
 module Cb
   module Clients
-    class Base
+    class DataList < Base
       class << self
-        def cb_client
-          @cb_client ||= Cb::Utils::Api.instance
+        # https://careerbuilder.readme.io/docs/consumerdatalistcountries
+        def get(args={})
+          uri = "#{ Cb.configuration.uri_data_list }/#{ args[:list] }"
+          cb_client.cb_get(uri, headers: headers(args), query: query_strings(args))
+        end
+
+        private
+
+        def query_strings(args)
+          qs = {}
+          qs[:countrycode] = args[:country_code] if args[:country_code]
+          qs
         end
 
         def headers(args)
           {
-             'Accept' => 'application/json',
-             'Authorization' => "Bearer #{ args[:oauth_token] }",
-             'Content-Type' => 'application/json'
+              'Accept' => 'application/json;version=2.0',
+              'Authorization' => "Bearer #{ args[:oauth_token] }",
+              'Content-Type' => 'application/json'
           }
         end
       end