+++ lib/cb/clients/user_profile.rb
@@ -8,20 +8,14 @@
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and limitations under the License.
+require_relative 'base'
 module Cb
   module Clients
-    class Base
+    class UserProfile < Base
       class << self
-        def cb_client
-          @cb_client ||= Cb::Utils::Api.instance
-        end
-
-        def headers(args)
-          {
-             'Accept' => 'application/json',
-             'Authorization' => "Bearer #{ args[:oauth_token] }",
-             'Content-Type' => 'application/json'
-          }
+        # https://careerbuilder.readme.io/docs/consumeruser-profile
+        def get(args={})
+          cb_client.cb_get(Cb.configuration.uri_user_profile, headers: headers(args))
         end
       end
     end