+++ lib/cb/utils/response_map.rb
@@ -40,22 +40,12 @@ module Cb
 
             Cb::Requests::ApplicationExternal::SubmitApplication => Cb::Responses::ApplicationExternal::SubmitApplication,
 
-            Cb::Requests::Category::Search => Cb::Responses::Category::Search,
-
             Cb::Requests::Company::Find => Cb::Responses::Company::Find,
-
-            Cb::Requests::CoverLetter::List => Cb::Responses::CoverLetter::List,
-            Cb::Requests::CoverLetter::Retrieve => Cb::Responses::CoverLetter::Retrieve,
-            Cb::Requests::CoverLetter::Update => Cb::Responses::CoverLetter::Update,
-            Cb::Requests::CoverLetter::Delete => Cb::Responses::CoverLetter::Delete,
-
-            Cb::Requests::Education::Get => Cb::Responses::Education::Get,
 
             Cb::Requests::EmailSubscription::Retrieve => Cb::Responses::EmailSubscription::Response,
             Cb::Requests::EmailSubscription::Modify => Cb::Responses::EmailSubscription::Response,
 
             Cb::Requests::Job::Report => Cb::Responses::Job::Report,
-
             Cb::Requests::JobSearch::Get => Cb::Responses::Job::Search,
 
             Cb::Requests::Recommendations::Resume => Cb::Responses::Recommendations,
@@ -64,17 +54,12 @@ module Cb
             Cb::Requests::Resumes::Put => Cb::Responses::Resume,
             Cb::Requests::Resumes::Delete => Cb::Responses::Resume,
             Cb::Requests::Resumes::Post => Cb::Responses::ResumeDocument,
-            Cb::Requests::Resumes::List => Cb::Responses::ResumeList,
-            Cb::Requests::Resumes::LanguageCodes => Cb::Responses::LanguageCodes,
 
             Cb::Requests::User::ChangePassword => Cb::Responses::User::ChangePassword,
             Cb::Requests::User::CheckExisting => Cb::Responses::User::CheckExisting,
             Cb::Requests::User::Delete => Cb::Responses::User::Delete,
             Cb::Requests::User::Retrieve => Cb::Responses::User::Retrieve,
-            Cb::Requests::User::TemporaryPassword => Cb::Responses::User::TemporaryPassword,
-
-            Cb::Requests::WorkStatus::List => Cb::Responses::WorkStatus::List,
-            Cb::Requests::State::Get => Cb::Responses::State
+            Cb::Requests::User::TemporaryPassword => Cb::Responses::User::TemporaryPassword
           }.merge response_hash_extension
         end
       end