+++ spec/spec_helper.rb
@@ -40,3 +40,4 @@ require 'support/convenience'
 include Cb::SpecSupport::Convenience
 
 require 'support/rspec/matchers/be_a_subclass_of'
+require 'support/common_client_stubs'